requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.6
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.4.26
click==8.2.1
deprecation==2.1.0
fastapi==0.115.12
frozenlist==1.6.0
gotrue==2.12.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
multidict==6.4.4
packaging==25.0
pluggy==1.6.0
postgrest==1.0.2
propcache==0.3.1
pydantic==2.11.5
pydantic_core==2.33.2
PyJWT==2.10.1
pytest==8.3.5
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
realtime==2.4.3
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
storage3==0.11.3
StrEnum==0.4.15
supabase==2.15.2
supafunc==0.9.4
typing-inspection==0.4.1
typing_extensions==4.13.2
uvicorn==0.34.3
websockets==14.2
yarl==1.20.0
